+++ src/HardcoreSystems.Mod/Modules/PowerGeneration/FuelThermalState.cs
@@ -15,6 +15,33 @@ namespace HardcoreSystems.Modules.PowerGeneration
         {
             get { return Samples.Count > 0; }
         }
+
+        public double OutputEnergyAssignedDtu { get; private set; }
+
+        public float MaximumTemperatureKelvin
+        {
+            get
+            {
+                var maximum = 0f;
+                for (var i = 0; i < Samples.Count; i++)
+                {
+                    if (Samples[i].TemperatureKelvin > maximum)
+                    {
+                        maximum = Samples[i].TemperatureKelvin;
+                    }
+                }
+
+                return maximum;
+            }
+        }
+
+        public void AddOutputEnergyAssigned(double energyDtu)
+        {
+            if (energyDtu > 0.0)
+            {
+                OutputEnergyAssignedDtu += energyDtu;
+            }
+        }
     }
 
     internal sealed class FuelThermalSample